add: utils/numbers/formatters file (moved funcs from transfer)
